backend/requirements.txt
# Production dependencies
# Pin exact versions for reproducible builds.

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Data validation ───────────────────────────────────────────────────────────
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.3.4
alembic==1.14.0

# ── Auth / Security ───────────────────────────────────────────────────────────
bcrypt==5.0.0
PyJWT==2.12.1

# ── AI providers ──────────────────────────────────────────────────────────────
anthropic==0.101.0
openai==1.109.1

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.4.0

# ── Task queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
